requirements.txt
adbc_driver_manager==1.4.0
adbc_driver_postgresql==1.4.0
adbc_driver_sqlite==1.4.0
anywidget==0.9.13
AppKit==0.2.8
argcomplete==3.5.3
asgiref==3.8.1
beautifulsoup4==4.13.3
boto3==1.36.23
botocore==1.36.23
brotli==1.1.0
brotlicffi==1.1.0.0
celery==5.4.0
chardet==5.2.0
checks==0.2
coloredlogs==15.0.1
configobj==5.0.9
ConfigParser==7.1.0
contextlib2==21.6.0
cryptography==44.0.1
cudf==0.6.1.post1
cycler==0.12.1
Cython==3.0.12
dash_leaflet==1.0.15
dash_testing_stub==0.0.2
dask==2025.2.0
dask_expr==2.0.0
diskcache==5.6.3
dl==0.1.0
docutils==0.21.2
duckdb==1.2.0
eventlet==0.39.0
fast_dash==0.2.10
fastparquet==2024.11.0
fire==0.7.0
flask_compress==1.17
Foundation==0.1.0a0.dev1
fsspec==2025.2.0
gevent==24.11.1
greenlet==3.1.1
h2==4.2.0
HTMLParser==0.0.2
hypothesis==6.126.0
ibis==3.3.0
importlib_resources==6.5.2
ini2toml==0.15
inotify==0.2.10
ipykernel==6.29.5
ipython==8.32.0
ipywidgets==8.1.5
jnius==1.1.0
js==1.0
keyring==25.6.0
lxml==5.3.1
matplotlib==3.10.0
mock==5.1.0
modin==0.32.0
mtrand==0.1
multiprocess==0.70.17
mypy==1.15.0
numexpr==2.10.2
odfpy==1.4.1
openpyxl==3.1.5
ordereddict==1.1
orjson==3.10.15
PasteScript==3.7.0
percy==2.0.2
pexpect==4.9.0
pickle5==0.0.12
Pillow==11.1.0
pkg1==0.0.3
polars==1.22.0
protobuf==5.29.3
psutil==7.0.0
PyInstaller==6.12.0
pyodide==0.0.2
pyOpenSSL==25.0.0
PyQt4==4.11.4
PyQt5==5.15.11
pyspark==3.5.4
python-dotenv==1.0.1
python_calamine==0.3.1
pyxlsb==1.0.10
qtpy==2.4.3
railroad==0.5.0
redis==5.2.1
s3fs==2025.2.0
scikit_learn==1.6.1
scipy==1.15.2
scipy_doctest==1.6
selenium==4.28.1
setproctitle==1.3.4
sets==0.3.2
simplejson==3.20.1
skimage==0.0
Sphinx==8.2.0
SQLAlchemy==2.0.38
sqlframe==3.22.0
statsmodels==0.14.4
tables==3.10.2
thread==2.0.5
threadpoolctl==3.5.0
tomli_w==1.2.0
tornado==6.4.2
traitlets==5.14.3
trove_classifiers==2025.2.18.16
Twisted==24.11.0
ukkonen==1.0.1
urllib3_secure_extra==0.1.0
vincent==0.4.4
watchdog==6.0.0
xarray==2025.1.2
xlrd==2.0.1
xlsxwriter==3.2.2
xmlrpclib==1.0.1
xx==3.3.2
zope==5.12
zstandard==0.23.0
gunicorn
pandas
dash-tools
